move mol_to_atoms to geometricanalysis.py
